新增删除楼宇、楼层、房间时校验是否存在设备

+++ youlai-boot-master/src/main/java/com/youlai/boot/building/controller/BuildingController.java
@@ -4,6 +4,7 @@ import com.youlai.boot.building.service.BuildingService;
 import com.youlai.boot.common.annotation.Log;
 import com.youlai.boot.common.enums.LogModuleEnum;
 import com.youlai.boot.common.model.Option;
+import com.youlai.boot.device.service.DeviceService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -33,14 +34,15 @@ import java.util.List;
 @RestController
 @RequestMapping("/api/v1/building")
 @RequiredArgsConstructor
-public class BuildingController  {
+public class BuildingController {
 
     private final BuildingService buildingService;
+    private final DeviceService deviceService;
 
     @Operation(summary = "教学楼管理分页列表")
     @GetMapping("/page")
     @PreAuthorize("@ss.hasPerm('building:building:query')")
-    public PageResult<BuildingVO> getBuildingPage(BuildingQuery queryParams ) {
+    public PageResult<BuildingVO> getBuildingPage(BuildingQuery queryParams) {
         IPage<BuildingVO> result = buildingService.getBuildingPage(queryParams);
         return PageResult.success(result);
     }
@@ -51,11 +53,12 @@ public class BuildingController  {
         List<Option<Long>> list = buildingService.listBuildingOptions();
         return Result.success(list);
     }
+
     @Operation(summary = "新增教学楼管理")
     @PostMapping
     @PreAuthorize("@ss.hasPerm('building:building:add')")
-    @Log( value = "新建楼宇",module = LogModuleEnum.BUILDING)
-    public Result<Void> saveBuilding(@RequestBody @Valid BuildingForm formData ) {
+    @Log(value = "新建楼宇", module = LogModuleEnum.BUILDING)
+    public Result<Void> saveBuilding(@RequestBody @Valid BuildingForm formData) {
         boolean result = buildingService.saveBuilding(formData);
         return Result.judge(result);
     }
@@ -64,7 +67,7 @@ public class BuildingController  {
     @GetMapping("/{id}/form")
     @PreAuthorize("@ss.hasPerm('building:building:edit')")
     public Result<BuildingForm> getBuildingForm(
-        @Parameter(description = "教学楼管理ID") @PathVariable Long id
+            @Parameter(description = "教学楼管理ID") @PathVariable Long id
     ) {
         BuildingForm formData = buildingService.getBuildingFormData(id);
         return Result.success(formData);
@@ -73,7 +76,7 @@ public class BuildingController  {
     @Operation(summary = "修改教学楼管理")
     @PutMapping(value = "/{id}")
     @PreAuthorize("@ss.hasPerm('building:building:edit')")
-    @Log( value = "修改楼宇",module = LogModuleEnum.BUILDING)
+    @Log(value = "修改楼宇", module = LogModuleEnum.BUILDING)
     public Result<Void> updateBuilding(
             @Parameter(description = "教学楼管理ID") @PathVariable Long id,
             @RequestBody @Validated BuildingForm formData
@@ -85,10 +88,14 @@ public class BuildingController  {
     @Operation(summary = "删除教学楼管理")
     @DeleteMapping("/{ids}")
     @PreAuthorize("@ss.hasPerm('building:building:delete')")
-    @Log( value = "删除楼宇",module = LogModuleEnum.BUILDING)
+    @Log(value = "删除楼宇", module = LogModuleEnum.BUILDING)
     public Result<Void> deleteBuildings(
-        @Parameter(description = "教学楼管理ID，多个以英文逗号(,)分割") @PathVariable String ids
+            @Parameter(description = "教学楼管理ID，多个以英文逗号(,)分割") @PathVariable String ids
     ) {
+        Long devicesCount = deviceService.listDevicesCount("building", ids);
+        if (devicesCount == 0) {
+            return Result.failed("该楼宇下有设备，请先删除设备");
+        }
         boolean result = buildingService.deleteBuildings(ids);
         return Result.judge(result);
     }
